playing around with UX, some of this will be refactored/removed
