+++ KlockWork-iOS/KlockWork-iOS/Pages/Explore/Explore.swift
@@ -51,7 +51,8 @@ struct Explore: View {
                         .frame(height: 45)
 
                     HStack(spacing: 8) {
-                        Text("Explore").font(.title2).padding([.leading, .trailing], 10).bold()
+                        DateStrip(date: self.date)
+                        Text("Explore").font(.title2).padding([.trailing], 10).bold()
                         Spacer()
                         CreateEntitiesButton()
                     }
