+++ KlockWork-iOS/KlockWork-iOS/Pages/Planning.swift
@@ -46,7 +46,6 @@ extension Planning {
         @EnvironmentObject private var state: AppState
         @State private var date: Date = Date()
         public let page: PageConfiguration.AppPage
-        
 
         var body: some View {
             HStack(alignment: .center, spacing: 0) {
@@ -57,7 +56,8 @@ extension Planning {
                         .frame(height: 45)
                     
                     HStack(spacing: 8) {
-                        Text("Planning").font(.title2).padding([.leading, .trailing], 10).bold()
+                        DateStrip(date: self.date)
+                        Text("Planning").font(.title2).padding([.trailing], 10).bold()
                         Spacer()
                         CreateEntitiesButton()
                     }
