+++ KlockWork-iOS/KlockWork-iOS/Pages/Today.swift
@@ -94,7 +94,8 @@ extension Today {
                         .frame(height: 45)
 
                     HStack(spacing: 8) {
-                        Text("Today").font(.title2).padding([.leading, .trailing], 10).bold()
+                        DateStrip(date: self.date)
+                        Text("Today").font(.title2).padding([.trailing], 10).bold()
                         Spacer()
                         CreateEntitiesButton(isViewModeSelectorVisible: true)
                     }
@@ -126,107 +127,6 @@ extension Today {
                     text: $text
                 )
                 .focused($focused)
-            }
-        }
-    }
-
-    struct AddButton: View {
-        typealias Entity = PageConfiguration.EntityType
-        @EnvironmentObject private var state: AppState
-        @State private var isPresented: Bool = false
-
-        var body: some View {
-            NavigationStack {
-                Menu("", systemImage: "plus") {
-                    NavigationLink {
-                        TaskDetail()
-                    } label: {
-                        Text(Entity.tasks.enSingular)
-                        Entity.tasks.icon
-                    }
-
-                    NavigationLink {
-                        NoteDetail.Sheet()
-                    } label: {
-                        Text(Entity.notes.enSingular)
-                        Entity.notes.icon
-                    }
-
-                    NavigationLink {
-                        PersonDetail()
-                    } label: {
-                        Text(Entity.people.enSingular)
-                        Entity.people.icon
-                    }
-
-                    NavigationLink {
-                        CompanyDetail()
-                    } label: {
-                        Text(Entity.companies.enSingular)
-                        Entity.companies.icon
-                    }
-
-                    NavigationLink {
-                        ProjectDetail()
-                    } label: {
-                        Text(Entity.projects.enSingular)
-                        Entity.projects.icon
-                    }
-
-                    NavigationLink {
-                        JobDetail()
-                    } label: {
-                        Text(Entity.jobs.enSingular)
-                        Entity.jobs.icon
-                    }
-                }
-            }
-            .font(.title2)
-            .foregroundStyle(self.state.theme.tint)
-        }
-    }
-
-    struct ViewModeSelector: View {
-        @EnvironmentObject private var state: AppState
-        @AppStorage("today.viewMode") private var storedVm: Int = 0
-        @State private var viewMode: ViewMode = .tabular
-
-        var body: some View {
-            VStack(alignment: .leading, spacing: 0) {
-                HStack(alignment: .center, spacing: 0) {
-                    Button {
-                        self.viewMode = .tabular
-                        self.storedVm = self.viewMode.id
-                    } label: {
-                        Image(systemName: "tablecells")
-                    }
-                    .disabled(self.storedVm == 0)
-                    .padding(5)
-                    .background(self.storedVm == 0 ? self.state.theme.tint : .black.opacity(0.1))
-                    .foregroundStyle(self.storedVm == 0 ? Theme.cPurple : self.state.theme.tint )
-                    .clipShape(.rect(topLeadingRadius: 6, bottomLeadingRadius: 6))
-
-                    Button {
-                        self.viewMode = .hierarchical
-                        self.storedVm = self.viewMode.id
-                    } label: {
-                        Image(systemName: "list.bullet")
-                    }
-                    .disabled(self.storedVm == 1)
-                    .padding(5)
-                    .background(self.storedVm == 1 ? self.state.theme.tint : .black.opacity(0.1))
-                    .foregroundStyle(self.storedVm == 1 ? Theme.cPurple : self.state.theme.tint)
-                    .clipShape(.rect(bottomTrailingRadius: 6, topTrailingRadius: 6))
-                }
-            }
-            .onAppear(perform: self.actionOnAppear)
-        }
-
-        /// Onload handler. Sets the viewMode to the stored value.
-        /// - Returns: Void
-        private func actionOnAppear() -> Void {
-            if let fromStored = ViewMode.by(id: storedVm) {
-                self.viewMode = fromStored
             }
         }
     }
